+++ common/src/configs/mod.rs
@@ -2,29 +2,33 @@ pub mod model_config;
 pub mod physics_config;
 pub mod player_config;
 pub mod scene_config;
+pub mod audio_config;
 
-use std::env;
 use serde::Serialize;
 use std::fs::File;
 use std::io::{Read, Write};
 use std::path::Path;
 use std::sync::{Arc, Mutex};
 use structopt::lazy_static;
-
 use crate::configs::model_config::ConfigModels;
 use crate::configs::scene_config::ConfigSceneGraph;
+use crate::configs::audio_config::ConfigAudioAssets;
 
 const MODELS_CONFIG_PATH: &str = "models.json";
 const SCENE_CONFIG_PATH: &str = "scene.json";
+const DISPLAY_CONFIG_PATH: &str = "display.json";
+const TEXTURE_CONFIG_PATH: &str = "tex.json";
+const AUDIO_CONFIG_PATH: &str = "audio.json";
 
 pub struct Config {
     pub models: ConfigModels,
     pub scene: ConfigSceneGraph,
+    pub audio: ConfigAudioAssets,
 }
 
 impl Config {
-    fn new(models: ConfigModels, scene: ConfigSceneGraph) -> Self {
-        Config { models, scene }
+    pub fn new(models: ConfigModels, scene: ConfigSceneGraph, audio: ConfigAudioAssets) -> Self {
+        Config { models, scene, audio }
     }
 }
 
@@ -34,15 +38,10 @@ lazy_static::lazy_static! {
 
 pub fn load_configuration() -> Result<(), Box<dyn std::error::Error>> {
     let models: ConfigModels = from_file(MODELS_CONFIG_PATH)?;
-
-    #[cfg(test)]
-    println!("{:?}", env::current_dir());
-    let scene: ConfigSceneGraph = from_file("../scene.json")?;
-
-    #[cfg(not(test))]
     let scene: ConfigSceneGraph = from_file(SCENE_CONFIG_PATH)?;
+    let audio: ConfigAudioAssets = from_file(AUDIO_CONFIG_PATH)?;
 
-    let config = Config::new(models, scene);
+    let config = Config::new(models, scene, audio);
     let mut instance = CONFIG_INSTANCE.lock().unwrap();
     *instance = Some(config);
